src/backend/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.txt
#
certifi==2017.11.5
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreschema==0.0.4
coverage==4.4.2
django-cors-headers==2.1.0
django-rest-swagger==2.1.2
django==2.0
djangorestframework==3.7.4
first==2.0.1
gunicorn==19.7.1
idna==2.6
importanize==0.6.1
itypes==1.1.0
jinja2==2.10
markupsafe==1.0
mysqlclient==1.3.12
openapi-codec==1.3.2
pathlib2==2.3.0
pip-tools==1.11.0
pytz==2017.3
requests==2.18.4
simplejson==3.13.2
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
